feat: space page migrated to Mantine
